+++ src/pages/CartList.jsx
@@ -1,18 +1,28 @@
-import { deleteCartItem, modifyCartItemCount } from 'actions/cart';
-import Button from 'components/@common/Button/styles';
-import CheckBox from 'components/@common/CheckBox';
-import * as CommonStyled from 'components/@common/CommonStyle/styles';
-import CartProducItem from 'components/CartProductItem';
-import Layout from 'components/Layout';
 import { useEffect, useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
+import { useCheckBox } from 'hooks';
+
+import CartProducItem from 'components/CartProductItem';
+import Layout from 'components/Layout';
+import Button from 'components/@common/Button/styles';
+import CheckBox from 'components/@common/CheckBox';
+import { snackbar } from 'actions/snackbar';
+import { deleteCartItem, modifyCartItemCount } from 'actions/cart';
 import { COLORS } from 'styles/theme';
+
+import * as CommonStyled from 'components/@common/CommonStyle/styles';
 import { CartListContainer, CartListReceiptContainer } from './styles';
 
 const CartList = () => {
   const { items: cartList } = useSelector((state) => state.cart);
-  const [selectAllChecked, setSelectAllChecked] = useState(false);
-  const [checkedList, setCheckedList] = useState([]);
+  const {
+    checkedList,
+    selectAllChecked,
+    handleChecked,
+    isChecked,
+    checkAllSelectButton,
+    deleteSelectedItem,
+  } = useCheckBox(cartList);
   const [totalPrice, setTotalPrice] = useState(0);
   const dispatch = useDispatch();
 
@@ -25,61 +35,12 @@ const CartList = () => {
     });
 
     setTotalPrice(calculateTotalPrice);
-
-    if (cartList.length > 0) {
-      if (checkedList.length >= cartList.length) {
-        setSelectAllChecked(true);
-      } else {
-        setSelectAllChecked(false);
-      }
-    }
   }, [cartList, checkedList]);
 
-  const handleChecked = (productId) => {
-    const prevState = [...checkedList];
-    const itemIndex = checkedList.findIndex((id) => id === productId);
-
-    if (itemIndex === -1) {
-      setCheckedList([...prevState, productId]);
-      return;
-    }
-
-    if (checkedList.length === 1) {
-      setCheckedList([]);
-      return;
-    }
-
-    setCheckedList(() => {
-      prevState.splice(itemIndex, 1);
-      return prevState;
-    });
-  };
-
-  const isChecked = (productId) => checkedList.findIndex((id) => id === productId) !== -1;
-
-  const checkAllSelectButton = () => {
-    if (cartList.length <= 0) {
-      setSelectAllChecked(!selectAllChecked);
-      return;
-    }
-
-    if (checkedList.length >= cartList.length) {
-      setCheckedList([]);
-      setSelectAllChecked(false);
-      return;
-    }
-
-    setCheckedList(cartList.map((item) => item.id));
-    setSelectAllChecked(true);
-  };
-
-  const deleteSelectedItem = () => {
-    if (checkedList <= 0) {
-      return;
-    }
-
+  const handleDeleteSelectedItem = () => {
     dispatch(deleteCartItem(checkedList));
-    setCheckedList([]);
+    dispatch(snackbar.pushMessageSnackbar('삭제했습니다! 다음에 구매해주세요 😊'));
+    deleteSelectedItem();
   };
 
   const handleItemCount = (productId, count) => {
@@ -103,7 +64,7 @@ const CartList = () => {
                 margin="0"
                 size="1rem"
                 weight="normal"
-                onClick={() => deleteSelectedItem()}
+                onClick={() => handleDeleteSelectedItem()}
               >
                 상품삭제
               </Button>
